adding Selector abstraction trait
